Fixes #352 validate yaml file and ignore invalid input instead of crashing
